# line reading slicing refactor