--- a/.travis.yml
+++ b/.travis.yml
@@ -1,4 +1,7 @@
 language: go
+go:
+  - 1.7.x
+  - 1.10.x
 sudo: false
 install:
   - go get github.com/stretchr/testify